@@ -12,42 +12,42 @@
 	},
 	"devDependencies": {
 		"@fontsource/inter": "^5.1.0",
-		"@iconify/json": "^2.2.250",
+		"@iconify/json": "^2.2.269",
 		"@iconify/tailwind": "^1.1.3",
 		"@iconify/utils": "^2.1.33",
 		"@neoconfetti/svelte": "^2.2.1",
-		"@sveltejs/adapter-static": "^3.0.5",
-		"@sveltejs/enhanced-img": "^0.3.8",
-		"@sveltejs/kit": "^2.5.28",
-		"@sveltejs/vite-plugin-svelte": "^3.1.2",
+		"@sveltejs/adapter-static": "^3.0.6",
+		"@sveltejs/enhanced-img": "^0.4.1",
+		"@sveltejs/kit": "^2.8.0",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@tailwindcss/forms": "^0.5.9",
 		"@types/aos": "^3.0.7",
 		"@types/eslint": "9.6.1",
-		"@typescript-eslint/eslint-plugin": "^8.6.0",
-		"@typescript-eslint/parser": "^8.6.0",
+		"@typescript-eslint/eslint-plugin": "^8.13.0",
+		"@typescript-eslint/parser": "^8.13.0",
 		"aos": "^2.3.4",
 		"autoprefixer": "^10.4.20",
 		"cssnano": "^7.0.6",
 		"cssnano-preset-default": "^7.0.6",
 		"cssnano-preset-lite": "^4.0.3",
-		"daisyui": "^4.12.10",
-		"embla-carousel-svelte": "^8.3.0",
-		"eslint": "^9.10.0",
+		"daisyui": "^4.12.14",
+		"embla-carousel-svelte": "^8.3.1",
+		"eslint": "^9.14.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.44.0",
+		"eslint-plugin-svelte": "^2.46.0",
 		"is-wsl2": "workspace:*",
 		"postcss": "^8.4.47",
 		"postcss-load-config": "^6.0.1",
 		"prettier": "^3.3.3",
-		"prettier-plugin-svelte": "^3.2.6",
-		"prettier-plugin-tailwindcss": "^0.6.6",
-		"svelte": "^4.2.19",
-		"svelte-body": "^1.4.0",
-		"svelte-check": "^4.0.2",
-		"tailwindcss": "^3.4.12",
-		"tslib": "^2.7.0",
-		"typescript": "^5.6.2",
-		"vite": "^5.4.6"
+		"prettier-plugin-svelte": "^3.2.7",
+		"prettier-plugin-tailwindcss": "^0.6.8",
+		"svelte": "^5.1.12",
+		"svelte-body": "^2.0.0",
+		"svelte-check": "^4.0.5",
+		"tailwindcss": "^3.4.14",
+		"tslib": "^2.8.1",
+		"typescript": "^5.6.3",
+		"vite": "^5.4.10"
 	},
 	"type": "module"
 }
