@@ -17,69 +17,68 @@
   },
   "devDependencies": {
     "@floating-ui/core": "^1.6.8",
-    "@floating-ui/dom": "^1.6.11",
-    "@iconify/json": "^2.2.250",
+    "@floating-ui/dom": "^1.6.12",
+    "@iconify/json": "^2.2.269",
     "@iconify/tailwind": "^1.1.3",
     "@iconify/utils": "^2.1.33",
-    "@inlang/paraglide-js": "1.11.2",
-    "@inlang/paraglide-sveltekit": "^0.11.0",
-    "@putout/minify": "^4.6.0",
-    "@skeletonlabs/skeleton": "^2.10.2",
+    "@inlang/paraglide-js": "1.11.3",
+    "@inlang/paraglide-sveltekit": "^0.11.5",
+    "@putout/minify": "^4.7.2",
+    "@skeletonlabs/skeleton": "^2.10.3",
     "@skeletonlabs/tw-plugin": "^0.4.0",
-    "@sveltejs/adapter-static": "^3.0.5",
-    "@sveltejs/kit": "^2.5.28",
-    "@sveltejs/vite-plugin-svelte": "^3.1.2",
+    "@sveltejs/adapter-static": "^3.0.6",
+    "@sveltejs/kit": "^2.8.0",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@tailwindcss/container-queries": "^0.1.1",
     "@tailwindcss/forms": "^0.5.9",
     "@tailwindcss/typography": "^0.5.15",
-    "@types/color": "^3.0.6",
+    "@types/color": "^4.2.0",
     "@types/firefox-webext-browser": "^120.0.4",
     "@types/randomcolor": "^0.5.9",
     "@types/webextension-polyfill": "^0.12.1",
-    "@typescript-eslint/eslint-plugin": "^8.6.0",
-    "@typescript-eslint/parser": "^8.6.0",
+    "@typescript-eslint/eslint-plugin": "^8.13.0",
+    "@typescript-eslint/parser": "^8.13.0",
     "autoprefixer": "^10.4.20",
     "browser-storage-polyfill": "^1.0.11",
-    "chart.js": "^4.4.4",
+    "chart.js": "^4.4.6",
     "color": "^4.2.3",
     "cssnano": "^7.0.6",
     "cssnano-preset-default": "^7.0.6",
     "cssnano-preset-lite": "^4.0.3",
     "date-fns": "^4.1.0",
-    "debounce": "^2.1.1",
-    "eslint": "^9.10.0",
+    "debounce": "^2.2.0",
+    "eslint": "^9.14.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-svelte": "^2.44.0",
+    "eslint-plugin-svelte": "^2.46.0",
     "fast-average-color": "^9.4.0",
     "fast-json-patch": "^3.1.1",
     "fflate": "^0.8.2",
-    "flat-ui-colors-json": "^0.1.0",
+    "flat-ui-colors-json": "^1.0.0",
     "hasha": "^6.0.0",
     "is-wsl2": "workspace:*",
-    "nanoid": "^5.0.7",
+    "nanoid": "^5.0.8",
     "openmeteo": "^1.1.4",
     "p-debounce": "^4.0.0",
     "postcss": "^8.4.47",
     "prettier": "^3.3.3",
-    "prettier-plugin-svelte": "^3.2.6",
-    "prettier-plugin-tailwindcss": "^0.6.6",
+    "prettier-plugin-svelte": "^3.2.7",
+    "prettier-plugin-tailwindcss": "^0.6.8",
     "randomcolor": "^0.6.2",
-    "svelte": "^4.2.19",
+    "svelte": "^5.1.12",
     "svelte-chartjs": "^3.1.5",
-    "svelte-check": "^4.0.2",
-    "svelte-fast-dimension": "^1.1.0",
+    "svelte-check": "^4.0.5",
     "svelte-floating-ui": "^1.5.9",
     "svelte-moveable": "^0.45.0",
     "svelte-selecto": "^1.26.3",
     "svelte-use-debounce": "^0.0.2",
     "svelte-virtual-scroll-list": "^1.3.0",
-    "tailwindcss": "^3.4.12",
+    "tailwindcss": "^3.4.14",
     "tiny-glob": "^0.2.9",
-    "tslib": "^2.7.0",
+    "tslib": "^2.8.1",
     "tslog": "^4.9.3",
-    "typescript": "^5.6.2",
+    "typescript": "^5.6.3",
     "vanilla-colorful": "^0.7.2",
-    "vite": "^5.4.6",
+    "vite": "^5.4.10",
     "vite-plugin-run-generator": "workspace:*",
     "vite-plugin-tailwind-purgecss": "^0.3.3",
     "web-ext": "^8.3.0",
