@@ -1,7 +1,7 @@
-<script lang="ts" context="module">
+<script lang="ts" module>
   import { Lazy } from '$lib/lazy';
   import { Opfs } from '$lib/opfs';
-  import { onMount } from 'svelte';
+  import { onMount, type Snippet } from 'svelte';
   import * as m from '$i18n/messages';
 
   export enum Constraints {
@@ -14,9 +14,13 @@
 </script>
 
 <script lang="ts">
-  export let constraint: Constraints;
-  let isSupported: boolean;
-  const { class: exClass, ...otherProps } = $$restProps;
+  let {
+    constraint,
+    class: exClass,
+    children,
+    ...otherProps
+  }: { constraint: Constraints; class: string; children: Snippet } = $props();
+  let isSupported: boolean | undefined = $state();
 
   onMount(async () => {
     isSupported = await ConstraintChecks[constraint].value;
@@ -24,7 +28,7 @@
 </script>
 
 {#if isSupported}
-  <slot />
+  {@render children()}
 {:else}
   <div class="bg-warning-100 border-warning-500 p-2 rounded {exClass || ''}" {...otherProps}>
     <p class="text-sm text-warning-900 text-center">{m.BrowserSupports_NotSupported()}</p>
