@@ -2,13 +2,13 @@
   import { opfsSrc } from '$actions/opfs-src';
   import type { WorkspaceInstance } from '$lib/workspace-instance';
 
-  export let workspaceInstance: WorkspaceInstance | undefined;
+  let { workspaceInstance }: { workspaceInstance: WorkspaceInstance } = $props();
 
-  $: icon16 = workspaceInstance?.favicon[16];
-  $: icon32 = workspaceInstance?.favicon[32];
-  $: iconIco = workspaceInstance?.favicon.ico;
+  let icon16 = $derived(workspaceInstance.favicon[16]);
+  let icon32 = $derived(workspaceInstance.favicon[32]);
+  let iconIco = $derived(workspaceInstance.favicon.ico);
 </script>
 
-<link rel="icon" type="image/png" sizes="32x32" use:opfsSrc={$icon32 || './favicon/favicon-32x32.png'} />
-<link rel="icon" type="image/png" sizes="16x16" use:opfsSrc={$icon16 || './favicon/favicon-16x16.png'} />
-<link rel="shortcut icon" use:opfsSrc={$iconIco || './favicon/favicon.ico'} />
+<link rel="icon" type="image/png" sizes="32x32" use:opfsSrc={icon32.value || './favicon/favicon-32x32.png'} />
+<link rel="icon" type="image/png" sizes="16x16" use:opfsSrc={icon16.value || './favicon/favicon-16x16.png'} />
+<link rel="shortcut icon" use:opfsSrc={iconIco.value || './favicon/favicon.ico'} />
