@@ -2,7 +2,7 @@
   import * as m from '$i18n/messages';
   import { ActiveFilters, AvailableWidgetFilters, type Filter } from '$stores/active-filters-store';
 
-  export let filter: Filter | undefined;
+  let { filter = $bindable() }: { filter?: Filter } = $props();
 
   function notifyFilterChanged(e: Event) {
     if (e.target instanceof HTMLSelectElement) {
@@ -17,7 +17,7 @@
   }
 </script>
 
-<select class="select" value={filter} on:change={notifyFilterChanged}>
+<select class="select" value={filter} onchange={notifyFilterChanged}>
   <option value={undefined}>{m.Widgets_Common_Settings_Filter_None()}</option>
   {#each AvailableWidgetFilters as wf}
     <option value={wf}>{wf}</option>
