@@ -1,12 +1,21 @@
 <script lang="ts">
   import { onDestroy } from 'svelte';
 
-  export let value: number | undefined;
-  export let min: number | undefined = undefined;
-  export let max: number | undefined = undefined;
-  export let step: number = 1;
-
-  const { placeholder, class: exClass } = $$restProps;
+  let {
+    value = $bindable(),
+    min = undefined,
+    max = undefined,
+    step = 1,
+    class: exClass,
+    placeholder,
+  }: {
+    value: number | undefined;
+    min?: number;
+    max?: number;
+    step?: number;
+    class?: string;
+    placeholder?: string;
+  } = $props();
 
   let timeout: ReturnType<typeof setTimeout> | undefined;
   let interval: ReturnType<typeof setInterval> | undefined;
@@ -62,11 +71,13 @@
 </script>
 
 <div class="input-group grid-cols-[auto_1fr_auto] {exClass || ''}">
-  <button class="variant-soft !pl-2 !pr-2" on:click={dec} on:mousedown={() => mousedown(dec)} on:mouseup={mouseup}>
+  <!-- svelte-ignore a11y_consider_explicit_label -->
+  <button class="variant-soft !pl-2 !pr-2" onclick={dec} onmousedown={() => mousedown(dec)} onmouseup={mouseup}>
     <span class="w-6 h-6 icon-[ic--baseline-minus]"></span>
   </button>
   <input class="no-spinner text-center !min-w-0" type="number" bind:value {min} {max} {placeholder} {step} />
-  <button class="variant-soft !pl-2 !pr-2" on:click={inc} on:mousedown={() => mousedown(inc)} on:mouseup={mouseup}>
+  <!-- svelte-ignore a11y_consider_explicit_label -->
+  <button class="variant-soft !pl-2 !pr-2" onclick={inc} onmousedown={() => mousedown(inc)} onmouseup={mouseup}>
     <span class="w-6 h-6 icon-[ic--baseline-plus]"></span>
   </button>
 </div>
