@@ -1,4 +1,4 @@
-<script context="module" lang="ts">
+<script module lang="ts">
   import * as m from '$i18n/messages';
 
   const TextTabId = 1;
@@ -21,29 +21,34 @@
   import { ClockFormat, type Settings } from './settings';
   import TextSettings from '$shared-components/text-settings.svelte';
   import BackgroundSettings from '$shared-components/background-settings.svelte';
+  import { onMount } from 'svelte';
 
-  export let settings: Settings;
-  export let tab: number;
-  export const tabs = Tabs;
+  let { settings, tab, tabs = $bindable() }: { settings: Settings; tab: number; tabs: object[] } = $props();
 
-  const { clockFormat, textColor, backgroundColor, backgroundBlur, font: fontSettings } = settings;
+  onMount(() => {
+    tabs = Tabs;
+  });
 </script>
 
 {#if tab === GeneralTabId}
-  <!-- svelte-ignore a11y-label-has-associated-control -->
+  <!-- svelte-ignore -->
   <label class="label mb-2">
     <span class="block">{m.Widgets_Clock_Settings_Format()}</span>
     <RadioGroup active="variant-filled-primary" hover="hover:variant-soft-primary">
-      <RadioItem bind:group={$clockFormat} name="justify" value={ClockFormat.TwelveHrs}>
+      <RadioItem bind:group={settings.clockFormat.value} name="justify" value={ClockFormat.TwelveHrs}>
         {m.Widgets_Clock_Settings_Format_12Hrs()}
       </RadioItem>
-      <RadioItem bind:group={$clockFormat} name="justify" value={ClockFormat.TwentyFourHrs}>
+      <RadioItem bind:group={settings.clockFormat.value} name="justify" value={ClockFormat.TwentyFourHrs}>
         {m.Widgets_Clock_Settings_Format_24Hrs()}
       </RadioItem>
     </RadioGroup>
   </label>
 {:else if tab === TextTabId}
-  <TextSettings font={fontSettings} bind:color={$textColor} shadow={settings.textShadow} stroke={settings.textStroke} />
+  <TextSettings
+    font={settings.font}
+    bind:color={settings.textColor.value}
+    shadow={settings.textShadow}
+    stroke={settings.textStroke} />
 {:else if tab === BackgroundTabId}
-  <BackgroundSettings bind:color={$backgroundColor} bind:blur={$backgroundBlur} />
+  <BackgroundSettings bind:color={settings.backgroundColor.value} bind:blur={settings.backgroundBlur.value} />
 {/if}
